Ensure case sub status set on finalise approve/refuse
